:recycle: (MessagingApps) Extract offer logic in new component
