+++ src/features/offer/components/shareMessagingOffer/MessagingApps.tsx
@@ -4,36 +4,27 @@ import styled from 'styled-components/native'
 
 import { InstalledMessagingApps } from 'features/offer/components/shareMessagingOffer/InstalledMessagingApps'
 import { MessagingAppContainer } from 'features/offer/components/shareMessagingOffer/MessagingAppContainer'
-import { useShareOffer } from 'features/share/helpers/useShareOffer'
 import { WebShareModal } from 'features/share/pages/WebShareModal'
-import { analytics } from 'libs/analytics'
+import { ShareContent } from 'libs/share'
 import { useModal } from 'ui/components/modals/useModal'
 import { ShareMessagingAppOther } from 'ui/components/ShareMessagingAppOther'
 import { Ul } from 'ui/components/Ul'
-import { getSpacing, Spacer, Typo } from 'ui/theme'
+import { Spacer, Typo, getSpacing } from 'ui/theme'
 import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
-type MessagingAppsProps = {
-  isEvent: boolean
-  offerId: number
+type Props = {
+  title: string
+  shareContent: ShareContent
+  share: () => void
+  messagingAppAnalytics: (social: Social | 'Other') => void
 }
 
-export const MessagingApps = ({ isEvent, offerId }: MessagingAppsProps) => {
-  const title = isEvent ? 'Vas-y en bande organisée\u00a0!' : 'Partage ce bon plan\u00a0!'
+export const MessagingApps = ({ title, shareContent, share, messagingAppAnalytics }: Props) => {
   const {
     visible: shareOfferModalVisible,
     showModal: showShareOfferModal,
     hideModal: hideShareOfferModal,
   } = useModal()
-
-  const { share, shareContent } = useShareOffer(offerId)
-
-  const messagingAppAnalytics = useCallback(
-    (social: Social | 'Other') => {
-      analytics.logShare({ type: 'Offer', id: offerId, from: 'offer', social })
-    },
-    [offerId]
-  )
 
   const onOtherPress = useCallback(() => {
     messagingAppAnalytics('Other')
@@ -41,7 +32,7 @@ export const MessagingApps = ({ isEvent, offerId }: MessagingAppsProps) => {
     showShareOfferModal()
   }, [messagingAppAnalytics, share, showShareOfferModal])
 
-  if (!shareContent || !shareContent.url) return null
+  if (!shareContent.url) return null
 
   return (
     <React.Fragment>
@@ -59,14 +50,14 @@ export const MessagingApps = ({ isEvent, offerId }: MessagingAppsProps) => {
         </StyledUl>
       </IconsWrapper>
       <Spacer.Column numberOfSpaces={4} />
-      {shareContent ? (
+      {!!shareContent && (
         <WebShareModal
           visible={shareOfferModalVisible}
           headerTitle="Partager l’offre"
           shareContent={shareContent}
           dismissModal={hideShareOfferModal}
         />
-      ) : null}
+      )}
     </React.Fragment>
   )
 }
